Large Commit: Refactor, and rework how touch is handeled.
Touch is now transfered from the dartboard directly to the dartpointer.  This way dartpointer handles all the touch events.  So the dartboard creates the new dart on touchstart.  Then all touch events on the board are transfered to the last dart placed.  Darts still have touch events assinged to handle indivudual darts being touched, however, this does not effect the touch events on the dartboard or other darts.